Updated: added argparse added logger separated to the different files
